src/requirements.txt
Flask==2.3.2
requests==2.31.0
pandas==2.0.3
google-cloud-bigquery==3.11.3
google-cloud-secret-manager==2.16.1